requirements.txt
# Core dependencies
python-telegram-bot==20.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
cryptography==41.0.5
alembic==1.12.1
apscheduler==3.10.4